+++ src/main/java/lotto/model/ErrorMessage.java
@@ -1,4 +1,4 @@
-package model;
+package lotto.model;
 
 public enum ErrorMessage {
     HEADER {
