+++ src/main/java/lotto/model/LottoStatus.java
@@ -1,4 +1,4 @@
-package model;
+package lotto.model;
 
 public enum LottoStatus {
     START(1),
